Create dockerignore and chnage Docker base image
dockerignore file was created so the docker build context is reduced. By
doing this we speed up docker build, and also make it more secure by not
importing secrets and/or keys.

Docker base image was changed to python:3.6-slim-stretch. With this
image, we have reduced the image size from around 1GB to 280MB.