Create new make direction and add requiremts file
We need a new make direction to build the lambda function with support
for ssh tunnel. By creating a new direction, user will be able to create
the lambda with and withou ssh tunel. I've split this in two because the
deployment sizes are diferent. Create a differen requirements file to
also achive the separation