Makefile (#1)
* Create dockerignore and chnage Docker base image

dockerignore file was created so the docker build context is reduced. By
doing this we speed up docker build, and also make it more secure by not
importing secrets and/or keys.

Docker base image was changed to python:3.6-slim-stretch. With this
image, we have reduced the image size from around 1GB to 280MB.

* Update dockerignore and gitignore

* Create new make direction and add requiremts file

We need a new make direction to build the lambda function with support
for ssh tunnel. By creating a new direction, user will be able to create
the lambda with and withou ssh tunel. I've split this in two because the
deployment sizes are diferent. Create a differen requirements file to
also achive the separation